Move mapper to marquez.core.common pkg
